register a new patient with the CUS for show

+++ src/main/java/com/tfg/medicalHistorybackend/models/jpa/PatientJPA.java
@@ -26,7 +26,7 @@ public class PatientJPA{
     private String healthInsurance;
 
     @Column(name = "member")
-    private int member;
+    private Long member;
 
     @Column(name = "status")
     private boolean status;
